requirements.txt
twitchio[starlette]
python-dotenv
websockets
pymongo
python-dotenv
requests
